.env not working but fixed up project restructure
