--- server/.env.local
@@ -1,4 +1,0 @@
-DB_HOST=localhost
-DB_USER=ggggg
-DB_PASSWORD=
-DB_NAME=
