+++ server/src/routes/items/delete.ts
