+++ server/src/routes/items/index.ts
@@ -1,8 +1,6 @@
-/*
 import { getTodoItems } from '../items/get'
-import { insertTodoItem } from './post'
+import { insertTodoItem } from '../items/post'
 import { updateTodoItem } from '../items/patch'
 import { deleteTodoItem } from '../items/delete'
 
 export { getTodoItems, insertTodoItem, updateTodoItem, deleteTodoItem }
-*/
